MAT-8972 fixing checkstyle issues
