+++ src/main/java/cms/gov/madie/measure/services/MeasureService.java
@@ -8,6 +8,7 @@ import cms.gov.madie.measure.repositories.MeasureRepository;
 import cms.gov.madie.measure.repositories.MeasureSetRepository;
 import cms.gov.madie.measure.repositories.TestCasePatchRepository;
 import cms.gov.madie.measure.resources.DuplicateKeyException;
+import cms.gov.madie.measure.utils.MeasureServiceUtil;
 import cms.gov.madie.measure.utils.MeasureUtil;
 import gov.cms.madie.models.access.AclOperation;
 import gov.cms.madie.models.access.AclSpecification;
@@ -29,7 +30,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.util.CollectionUtils;
 
-import java.text.SimpleDateFormat;
 import java.time.Instant;
 import java.time.LocalTime;
 import java.time.ZoneOffset;
@@ -59,7 +59,7 @@ public class MeasureService {
           "No measure set exists for measure set ID " + measureSetId);
     }
 
-    verifyMeasureSetAuthorization(
+    MeasureServiceUtil.verifyMeasureSetAuthorization(
         username,
         "MeasureSet",
         measureSetId,
@@ -97,26 +97,8 @@ public class MeasureService {
       throw new InvalidMeasureStateException(
           "No measure set exists for measure with ID " + measure.getId());
     }
-    verifyMeasureSetAuthorization(username, "Measure", measure.getId(), roles, measureSet);
-  }
-
-  private void verifyMeasureSetAuthorization(
-      String username,
-      String target,
-      String targetId,
-      List<RoleEnum> roles,
-      MeasureSet measureSet) {
-    List<RoleEnum> allowedRoles = roles == null ? List.of() : roles;
-    if (!measureSet.getOwner().equalsIgnoreCase(username)
-        && (CollectionUtils.isEmpty(measureSet.getAcls())
-            || measureSet.getAcls().stream()
-                .noneMatch(
-                    acl ->
-                        acl.getUserId().equalsIgnoreCase(username)
-                            && acl.getRoles().stream().anyMatch(allowedRoles::contains)))) {
-      log.error("User {} is not authorized for {} with target ID {}", username, target, targetId);
-      throw new UnauthorizedException(target, targetId, username);
-    }
+    MeasureServiceUtil.verifyMeasureSetAuthorization(
+        username, "Measure", measure.getId(), roles, measureSet);
   }
 
   // TODO: start replacing usage of measureRepository.findById with this method
@@ -142,7 +124,7 @@ public class MeasureService {
       Measure measure, final String username, String accessToken, boolean addDefaultCQL) {
     log.info("User [{}] is attempting to create a new measure", username);
     checkDuplicateCqlLibraryName(measure.getCqlLibraryName());
-    validateMeasurementPeriod(
+    MeasureServiceUtil.validateMeasurementPeriod(
         measure.getMeasurementPeriodStart(), measure.getMeasurementPeriodEnd());
     updateMeasurementPeriods(measure);
     Measure measureCopy = measure.toBuilder().build();
@@ -288,15 +270,15 @@ public class MeasureService {
     }
 
     if (measureUtil.isMeasurementPeriodChanged(updatingMeasure, existingMeasure)) {
-      validateMeasurementPeriod(
+      MeasureServiceUtil.validateMeasurementPeriod(
           updatingMeasure.getMeasurementPeriodStart(), updatingMeasure.getMeasurementPeriodEnd());
       updateMeasurementPeriods(updatingMeasure);
     }
 
-    updateReferenceId(updatingMeasure.getMeasureMetaData());
+    updateReferences(updatingMeasure.getMeasureMetaData());
 
     if (!ModelType.QDM_5_6.getValue().equalsIgnoreCase(updatingMeasure.getModel())) {
-      updateMeasureDefinitionId(updatingMeasure.getMeasureMetaData());
+      updateMeasureDefinitions(updatingMeasure.getMeasureMetaData());
     }
 
     Measure outputMeasure = updatingMeasure;
@@ -394,29 +376,6 @@ public class MeasureService {
     }
   }
 
-  public void validateMeasurementPeriod(Date measurementPeriodStart, Date measurementPeriodEnd) {
-    if (measurementPeriodStart == null || measurementPeriodEnd == null) {
-      throw new InvalidMeasurementPeriodException(
-          "Measurement period date is required and must be valid");
-    }
-    SimpleDateFormat checkYear = new SimpleDateFormat("yyyy");
-    int checkMeasurementPeriodStart = Integer.parseInt(checkYear.format(measurementPeriodStart));
-    int checkMeasurementPeriodEnd = Integer.parseInt(checkYear.format(measurementPeriodEnd));
-
-    if (1900 > checkMeasurementPeriodStart
-        || checkMeasurementPeriodStart > 2099
-        || 1900 > checkMeasurementPeriodEnd
-        || checkMeasurementPeriodEnd > 2099) {
-      throw new InvalidMeasurementPeriodException(
-          "Measurement periods should be between the years 1900 and 2099.");
-    }
-
-    if (measurementPeriodEnd.compareTo(measurementPeriodStart) < 1) {
-      throw new InvalidMeasurementPeriodException(
-          "Measurement period end date should be greater than measurement period start date.");
-    }
-  }
-
   public void checkDeletionCredentials(String username, String createdBy) {
     if (!username.equalsIgnoreCase(createdBy)) {
       throw new InvalidDeletionCredentialsException(username);
@@ -434,16 +393,6 @@ public class MeasureService {
       return measure.toBuilder().elmJson(elmJson.getJson()).elmXml(elmJson.getXml()).build();
     }
     return measure;
-  }
-
-  public void checkVersionIdChanged(String changedVersionId, String originalVersionId) {
-    if (StringUtils.isBlank(changedVersionId) && !StringUtils.isBlank(originalVersionId)) {
-      throw new InvalidVersionIdException(changedVersionId);
-    } else if (!StringUtils.isBlank(changedVersionId)
-        && !StringUtils.isBlank(originalVersionId)
-        && !changedVersionId.equalsIgnoreCase(originalVersionId)) {
-      throw new InvalidVersionIdException(changedVersionId);
-    }
   }
 
   public List<AclSpecification> updateAccessControlList(
@@ -714,24 +663,18 @@ public class MeasureService {
         username, pageReq, searchCriteria, ownershipTypes, invocationSource);
   }
 
-  protected void updateReferenceId(MeasureMetaData metaData) {
+  protected void updateReferences(MeasureMetaData metaData) {
     if (metaData != null && !CollectionUtils.isEmpty(metaData.getReferences())) {
       List<Reference> references =
-          (List<Reference>)
-              metaData.getReferences().stream()
-                  .map(reference -> updateReference(reference))
-                  .toList();
+          metaData.getReferences().stream().map(this::updateReference).toList();
       metaData.setReferences(references);
     }
   }
 
-  protected void updateMeasureDefinitionId(MeasureMetaData metaData) {
+  protected void updateMeasureDefinitions(MeasureMetaData metaData) {
     if (metaData != null && !CollectionUtils.isEmpty(metaData.getMeasureDefinitions())) {
       List<MeasureDefinition> definitions =
-          (List<MeasureDefinition>)
-              metaData.getMeasureDefinitions().stream()
-                  .map(definition -> updateMeasureDefinition(definition))
-                  .toList();
+          metaData.getMeasureDefinitions().stream().map(this::updateMeasureDefinition).toList();
       metaData.setMeasureDefinitions(definitions);
     }
   }
